+++ DataAccess/Repositories/RepositoryMaquinas.cs
@@ -37,7 +37,7 @@ namespace DataAccess.Repositories
             }
         }
 
-        public void Remove(int id)
+        public void Remove(decimal id)
         {
             using (var context = new InteleSysEntities())
             {
